+++ CHANGELOG.md
@@ -3,3 +3,7 @@ Changelog
 
 - 1.0
     - Initial public release as package for MODX Revolution
+- 1.0.1
+    - Add: param ignoreFields to db2Formit
+    - Change: paramname/fieldname of db2Formit are compared with the REQUEST array instead of GET array
+    - Bugfix: wrong parameter names in db2Formit code (now paramname/fieldname as documented)